refactor(test): extract reusable provider cassette support
