+++ Cargo.toml
@@ -65,6 +65,7 @@ members = [
 default-members = [
     ".",
     "crates/rig-core",
+    "crates/rig-cassette",
     "crates/rig-effect-log",
     "crates/rig-agent",
     "crates/rig-ecs",
@@ -316,6 +317,8 @@ rig-vertexai = { path = "crates/rig-vertexai", version = "0.42.0", optional = tr
 rig-tungstenite = { path = "crates/rig-tungstenite", version = "0.42.0", optional = true }
 
 [dev-dependencies]
+# The repository census checks Bedrock binary fixtures even in default suites.
+rig-cassette = { path = "crates/rig-cassette", features = ["bedrock"] }
 rig-ecs = { path = "crates/rig-ecs" }
 bevy_app = { workspace = true }
 bevy_ecs = { workspace = true }
